+++ CHANGELOG.md
@@ -1,6 +1,16 @@
 # Change Log
 All notable changes to this project will be documented in this file. 
 
+## 0.2.2 - 2016-09-24
+### Changed
+
+* Straighten out core and config namespaces for API
+* Added :old? and :format to options map for lein-zprint
+
+### Removed
+
+### Fixed
+
 ## 0.2.1 - 2016-09-22
 ### Changed